aula 15 / 08 - web frontend

+++ poo/aula-08-13/Calc.java
@@ -21,7 +21,7 @@ public class Calc {
         return x * y;
     }
     
-    public static int div(int x, int y) {
+    public static double div(int x, int y) {
         return x / y;
     }
 }
